refactor: 불필요한 JdbcDao 삭제
