--- app/src/main/java/com/techcourse/dao/JdbcTemplateUserHistoryDao.java
@@ -1,28 +1,0 @@
-package com.techcourse.dao;
-
-import com.interface21.jdbc.core.JdbcTemplate;
-import com.techcourse.domain.UserHistory;
-
-public class JdbcTemplateUserHistoryDao implements UserHistoryDao {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public JdbcTemplateUserHistoryDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Override
-    public void log(UserHistory userHistory) {
-        String sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
-
-        jdbcTemplate.update(
-                sql,
-                userHistory.getUserId(),
-                userHistory.getAccount(),
-                userHistory.getPassword(),
-                userHistory.getEmail(),
-                userHistory.getCreatedAt(),
-                userHistory.getCreatedBy()
-        );
-    }
-}
