--- app/src/main/java/com/techcourse/dao/JdbcUserHistoryDao.java
@@ -1,42 +1,0 @@
-package com.techcourse.dao;
-
-import com.interface21.dao.DataAccessException;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class JdbcUserHistoryDao implements UserHistoryDao {
-
-    private static final Logger log = LoggerFactory.getLogger(JdbcUserHistoryDao.class);
-
-    private final DataSource dataSource;
-
-    public JdbcUserHistoryDao(final DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    @Override
-    public void log(final UserHistory userHistory) {
-        String sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
-
-        try (Connection conn = dataSource.getConnection();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-            log.debug("query : {}", sql);
-
-            pstmt.setLong(1, userHistory.getUserId());
-            pstmt.setString(2, userHistory.getAccount());
-            pstmt.setString(3, userHistory.getPassword());
-            pstmt.setString(4, userHistory.getEmail());
-            pstmt.setObject(5, userHistory.getCreatedAt());
-            pstmt.setString(6, userHistory.getCreatedBy());
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
-    }
-}
