+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,17 +1,43 @@
 package com.techcourse.dao;
 
+import com.interface21.jdbc.core.JdbcTemplate;
+import com.interface21.jdbc.core.RowMapper;
+import com.techcourse.dao.mapper.UserMapper;
 import com.techcourse.domain.User;
 import java.util.List;
 
-public interface UserDao {
+public class UserDao {
 
-    void insert(final User user);
+    private static final RowMapper<User> ROW_MAPPER = new UserMapper();
 
-    void update(final User user);
+    private final JdbcTemplate jdbcTemplate;
 
-    List<User> findAll();
+    public UserDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
 
-    User findById(final Long id);
+    public void insert(User user) {
+        String sql = "INSERT INTO users (account, password, email) VALUES (?, ?, ?)";
+        jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail());
+    }
 
-    User findByAccount(final String account);
+    public void update(User user) {
+        String sql = "UPDATE users SET account=?, password=?, email=? WHERE id=?";
+        jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail(), user.getId());
+    }
+
+    public List<User> findAll() {
+        String sql = "SELECT id, account, password, email FROM users";
+        return jdbcTemplate.queryForList(sql, ROW_MAPPER);
+    }
+
+    public User findById(Long id) {
+        String sql = "SELECT id, account, password, email FROM users WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, ROW_MAPPER, id);
+    }
+
+    public User findByAccount(String account) {
+        String sql = "SELECT id, account, password, email FROM users WHERE account=?";
+        return jdbcTemplate.queryForObject(sql, ROW_MAPPER, account);
+    }
 }
