+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -1,8 +1,27 @@
 package com.techcourse.dao;
 
+import com.interface21.jdbc.core.JdbcTemplate;
 import com.techcourse.domain.UserHistory;
 
-public interface UserHistoryDao {
+public class UserHistoryDao {
 
-    void log(final UserHistory userHistory);
+    private final JdbcTemplate jdbcTemplate;
+
+    public UserHistoryDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public void log(UserHistory userHistory) {
+        String sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
+
+        jdbcTemplate.update(
+                sql,
+                userHistory.getUserId(),
+                userHistory.getAccount(),
+                userHistory.getPassword(),
+                userHistory.getEmail(),
+                userHistory.getCreatedAt(),
+                userHistory.getCreatedBy()
+        );
+    }
 }
