--- app/src/test/java/com/techcourse/dao/JdbcTemplateUserDaoTest.java
@@ -1,76 +1,0 @@
-package com.techcourse.dao;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.interface21.jdbc.core.JdbcTemplate;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.domain.User;
-import com.techcourse.support.jdbc.init.DatabasePopulatorUtils;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class JdbcTemplateUserDaoTest {
-
-    private UserDao userDao;
-    private JdbcTemplate jdbcTemplate;
-
-    @BeforeEach
-    void setup() {
-        DatabasePopulatorUtils.execute(DataSourceConfig.getInstance());
-        jdbcTemplate = new JdbcTemplate(DataSourceConfig.getInstance());
-        userDao = new JdbcTemplateUserDao(jdbcTemplate);
-        final var user = new User("gugu", "password", "hkkang@woowahan.com");
-        userDao.insert(user);
-    }
-
-    @DisplayName("모든 유저 정보를 조회한다.")
-    @Test
-    void findAll() {
-        final var users = userDao.findAll();
-
-        assertThat(users).isNotEmpty();
-    }
-
-    @DisplayName("아이디로 유저 정보를 조회한다.")
-    @Test
-    void findById() {
-        final var user = userDao.findById(1L);
-
-        assertThat(user.getAccount()).isEqualTo("gugu");
-    }
-
-    @DisplayName("계정으로 유저 정보를 조회한다.")
-    @Test
-    void findByAccount() {
-        final var account = "gugu";
-        final var user = userDao.findByAccount(account);
-
-        assertThat(user.getAccount()).isEqualTo(account);
-    }
-
-    @DisplayName("유저를 추가한다.")
-    @Test
-    void insert() {
-        final var account = "insert-gugu";
-        final var user = new User(account, "password", "hkkang@woowahan.com");
-        userDao.insert(user);
-
-        final var actual = userDao.findById(2L);
-
-        assertThat(actual.getAccount()).isEqualTo(account);
-    }
-
-    @DisplayName("유저 정보를 업데이트 한다.")
-    @Test
-    void update() {
-        final var newPassword = "password99";
-        final var user = userDao.findById(1L);
-        user.changePassword(newPassword);
-
-        userDao.update(user);
-        final var actual = userDao.findById(1L);
-
-        assertThat(actual.getPassword()).isEqualTo(newPassword);
-    }
-}
