+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -2,10 +2,10 @@ package com.techcourse.service;
 
 import com.interface21.dao.DataAccessException;
 import com.interface21.jdbc.core.JdbcTemplate;
-import com.techcourse.dao.JdbcTemplateUserHistoryDao;
+import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.UserHistory;
 
-public class MockUserHistoryDao extends JdbcTemplateUserHistoryDao {
+public class MockUserHistoryDao extends UserHistoryDao {
 
     public MockUserHistoryDao(final JdbcTemplate jdbcTemplate) {
         super(jdbcTemplate);
